+++ src/main/java/insane96mcp/survivalreimagined/module/mining/block/AbstractMultiBlockFurnace.java
@@ -1,7 +1,9 @@
 package insane96mcp.survivalreimagined.module.mining.block;
 
+import insane96mcp.survivalreimagined.SurvivalReimagined;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionHand;
@@ -10,7 +12,9 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.entity.AbstractFurnaceBlockEntity;
@@ -24,10 +28,15 @@ import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.block.state.properties.DirectionProperty;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 public abstract class AbstractMultiBlockFurnace extends BaseEntityBlock {
     public static final DirectionProperty FACING = HorizontalDirectionalBlock.FACING;
     public static final BooleanProperty LIT = BlockStateProperties.LIT;
+
+    public static final String MULTI_BLOCK_TOOLTIP = SurvivalReimagined.MOD_ID + ".multi_block_structure";
 
     public AbstractMultiBlockFurnace(Properties pProperties) {
         super(pProperties);
@@ -138,5 +147,10 @@ public abstract class AbstractMultiBlockFurnace extends BaseEntityBlock {
     @javax.annotation.Nullable
     protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level pLevel, BlockEntityType<T> pServerType, BlockEntityType<? extends AbstractMultiBlockFurnaceBlockEntity> pClientType) {
         return pLevel.isClientSide ? null : createTickerHelper(pServerType, pClientType, AbstractMultiBlockFurnaceBlockEntity::serverTick);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack pStack, @Nullable BlockGetter pLevel, List<Component> pTooltip, TooltipFlag pFlag) {
+        pTooltip.add(Component.translatable(MULTI_BLOCK_TOOLTIP));
     }
 }