+++ src/main/java/insane96mcp/survivalreimagined/module/mining/block/MultiBlockBlastFurnaceBlockEntity.java
@@ -1,6 +1,6 @@
 package insane96mcp.survivalreimagined.module.mining.block;
 
-import insane96mcp.survivalreimagined.module.mining.feature.SoulSteel;
+import insane96mcp.survivalreimagined.module.mining.feature.MultiBlockFurnaces;
 import insane96mcp.survivalreimagined.module.mining.inventory.MultiBlockBlastFurnaceMenu;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
@@ -11,7 +11,7 @@ import net.minecraft.world.level.block.state.BlockState;
 
 public class MultiBlockBlastFurnaceBlockEntity extends AbstractMultiBlockFurnaceBlockEntity {
     public MultiBlockBlastFurnaceBlockEntity(BlockPos pPos, BlockState pBlockState) {
-        super(SoulSteel.BLAST_FURNACE_BLOCK_ENTITY_TYPE.get(), pPos, pBlockState, SoulSteel.RECIPE_TYPE.get());
+        super(MultiBlockFurnaces.BLAST_FURNACE_BLOCK_ENTITY_TYPE.get(), pPos, pBlockState, MultiBlockFurnaces.BLASTING_RECIPE_TYPE.get());
     }
 
     @Override