+++ src/main/java/insane96mcp/survivalreimagined/module/mining/client/AbstractMultiBlockFurnaceScreen.java
@@ -27,8 +27,8 @@ public abstract class AbstractMultiBlockFurnaceScreen<T extends AbstractMultiBlo
     public void init() {
         super.init();
         this.widthTooNarrow = this.width < 379;
-        /*this.recipeBookComponent.init(this.width, this.height, this.minecraft, this.widthTooNarrow, this.menu);
-        this.leftPos = this.recipeBookComponent.updateScreenPosition(this.width, this.imageWidth);
+        this.recipeBookComponent.init(this.width, this.height, this.minecraft, this.widthTooNarrow, this.menu);
+        /*this.leftPos = this.recipeBookComponent.updateScreenPosition(this.width, this.imageWidth);
         this.addRenderableWidget(new ImageButton(this.leftPos + 20, this.height / 2 - 31, 20, 18, 0, 0, 19, RECIPE_BUTTON_LOCATION, (button) -> {
             this.recipeBookComponent.toggleVisibility();
             this.leftPos = this.recipeBookComponent.updateScreenPosition(this.width, this.imageWidth);